Moved shared types to store/types dir.
